+++ Code.playground/Sources/NARS/Utility/Judgement+.swift
@@ -1,7 +1,7 @@
 //import Dispatch
 
 extension Judgement {
-    static func updateTimestamp(_ j: Judgement, _ timeProvider: () -> UInt64) -> Judgement {
+    static func updateTimestamp(_ j: Judgement, _ timeProvider: () -> UInt32) -> Judgement {
         var j = j
         let now = timeProvider()
         if j.derivationPath.count == 1 { // also update derivationPath
